% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countrycode}
\alias{countrycode}
\title{Country iso3c and name mapping for PISA OECD countries participants.}
\format{
A tibble of the following variables
\itemize{
    \item \code{country}: Country 3 character code. Note that some regions/territories are coded as country for ease of input. Character.
    \item \code{country_name}: Country name. Note that some regions/territories are coded as country for ease of input. Character.
    }
}
\description{
A dataset containing mapping of the country ISO code to the country names.
More information on participating countries can be found at
\url{https://www.oecd.org/pisa/aboutpisa/pisa-participants.htm}.
}
