% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{school}
\alias{school}
\title{School data from 2000-2018 PISA OECD data}
\format{A tibble of the following variables
\itemize{
    \item student_id The student identification number
    \item stu_wgt The final survey weight score for the student score
    \item country_iso3c Country 3 character code
    \item school_id The unique school identification number
    \item funding_gov Percentage of total funding for school year from: Government
    \item funding_fees Percentage of total funding for school year from: Student fees or school charges paid by parents
    \item funding_donations Percentage of total funding for school year from: Benefactors, donations, bequests, sponsorships, parent fundraising
    \item enrollment_boys Number of boys
    \item enrollment_girls Number of girls
    \item student_teacher_ratio Student-Teacher ratio
    \item public_private Public or private school
    \item staff_shortage The unique school identification number
    \item sch_wgt Adjusted school base weight
    \item school_size The school size
    }}
\description{
A dataset containing school weight and other information
from the triennial testing of 15 year olds around
the globe. Original data available from
 \url{https://www.oecd.org/pisa/data/}.
}
