% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{school}
\alias{school}
\title{School data from 2000-2018 PISA OECD data}
\format{
A tibble of the following variables
\itemize{
    \item \code{year}: Year of the PISA data. Factor.
    \item \code{country_iso3c}: Country 3 character code. Factor.
    \item \code{school_id}: The school identification number, unique for each country and year combination. Factor.
    \item \code{fund_gov}: Percentage of total funding for school year from government. Numeric.
    \item \code{fund_fees}: Percentage of total funding for school year from student fees or school charges paid by parents. Numeric.
    \item \code{fund_donation}: Percentage of total funding for school year from
    benefactors, donations, bequests, sponsorships, parent fundraising. Numeric.
    \item \code{enrol_boys}: Number of boys in the school. Numeric.
    \item \code{enrol_girls}: Number of girls in the schoo. Numeric.
    \item \code{stratio}: Student-Teacher ratio. Numeric.
    \item \code{public_private}: Is the school a public or private school. Factor.
    \item \code{staff_shortage}: Shortage of staff. Numeric.
    \item \code{sch_wgt}: The final survey weight score for the schools. Numeric.
    \item \code{school_size}: The school size. Numeric.
    }
}
\description{
A dataset containing school weight and other information
from the triennial testing of 15 year olds around
the globe. Original data available from
 \url{https://www.oecd.org/pisa/data/}.
}
