% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{student}
\alias{student}
\title{Student data from 2000-2018 PISA OECD data}
\format{A tibble of the following variables
\itemize{
    \item country_iso3c Country 3 character code
    \item school_id The unique school identification number
    \item student_id The student identification number
    \item mother_educ Highest level of mother's education - 1 to 5
    \item father_educ Highest level of father's education - 1 to 5
    \item gender recorded gender - 1 female or 2 male or missing
    \item computer Possession of computer - 1 yes, 2 no, other values are coding missing values
    \item internet Access to internet - 1 yes, 2 no or missing
    \item math Simulated score in mathematics
    \item read Simulated score in reading
    \item science Simulated score in science
    \item stu_wgt The final survey weight score for the student score
    \item desk Possession of  desk to study at - 1 yes, 2 no or missing
    \item room Possession of a room of your own - 1 yes, 2 no or missing
    \item dishwasher Possession of a dishwasher - 1 yes, 2 no or missing
    \item television Number of televisions - 1 yes, 2 no or missing
    \item computer_n Number of computers - 1 yes, 2 no or missing
    \item car Number of cars - 1 yes, 2 no or missing
    \item book Number of books - 1 yes, 2 no or missing
    \item wealth Family wealth
    \item escs Index of economic, social and cultural status
    }}
\description{
A dataset containing scores and other information
from the triennial testing of 15 year olds around
the globe. Original data available from
 \url{https://www.oecd.org/pisa/data/}.
}
